+++ sherpa-onnx/java-api/src/com/k2fsa/sherpa/onnx/OnlineCtcFstDecoderConfig.java
@@ -31,7 +31,7 @@ public class OnlineCtcFstDecoderConfig {
             return new OnlineCtcFstDecoderConfig(this);
         }
 
-        public Builder setGraph(String model) {
+        public Builder setGraph(String graph) {
             this.graph = graph;
             return this;
         }